Added the function in functions.php, removed it from post.php
